made generic Int

+++ src/readquartetdata.jl
@@ -167,7 +167,7 @@ function readtableCF!(df::DataFrames.DataFrame, co::Vector{Int}; mergerows=false
     end
     d = DataCF(quartets)
     if "ngenes" in names(df)
-        d.numTrees = Int64(minimum(df.ngenes))
+        d.numTrees = Int(minimum(df.ngenes))
     end
     if(!isempty(repSpecies))
         d.repSpecies = repSpecies
